A number of changes intended to address https://wind-dev.nrel.gov/bugzilla/show_bug.cgi?id=430
In order to make sure the generated code still compiles and links, also "stubbed out" a bunch of places in the FAST branch here.  This is old code that's being superceded by changes Bonnie is making.  Most of these stub-outings had to do with references or definitions coming from Aerodyn (the old form before the new fmwk.  So basically right now the whole thing compiles and links from a make superclean and a make in the FAST Source dirctory.  

git-svn-id: https://windsvn2.nrel.gov/FAST/branches/CSC/FAST/Registry/Source@88 10acb478-4768-415a-8850-bacdb5912d4d