John,
Go ahead and check in your fix. I don’t have any objections… I was actually going to talk to you about this sometime, but I figured it was a minor issue at this stage.  I will change my file, and let anyone else who is using subtypes know (I don’t think anyone else is, yet).

Thanks for looking into this so quickly!

Bonnie

From: Michalakes, John 
Sent: Sunday, December 23, 2012 7:11 AM
To: Jonkman, Bonnie
Subject: RE: [Bug 430] [FAST registry] New: Types with subtypes ignore the subtype in the pack/unpack routines.

Okay, I figured it out.   On windows it wasn’t properly including a header file that defined those functions (strcpy, strncpy, and index).  But please read on… 

I haven’t checked in a fix yet, because I ran across another problem that I wanted to run past you first before I check in a fix.  Changing this may cause some problems for applications codes that are using the current version of the registry.

Looking through the registry file you attached to the bug report, I saw you had already noticed and adjusted for the problem, which is this:  any derived type that a user defines in the registry file gets the nickname of the module prepended to it.  For example, CoordSys becomes StrD_CoordSys .  So that one has to magically intuit (as you did) that to use the type as a subtype in another type, the StrD has to be prepended to the type name in the registry.  Sorry for that, that wasn’t intended. You were too nice in just adopting a severe limp to make a badly tailored suit fit correctly.  (She: “Oh, that poor man!”,  He: “Oh my yes, dear, let’s not stare.  But what a nice suit!”).  

The updates I have and that I’m ready to commit will fix that.  Only the derived types that are explicitly part of the FAST interface and listed in that table of the Programmer’s Handbook will get the nickname prepended.  These would never be used as a subtype  in another type (and if they are, one would have to prepend the nickname explicitly).  But any other user-defined type such as your CoordSys type would not have the nickname prepended.   It would just live as itself.  There’s very little chance of namespace collisions from this decision, and even for the rare instance, there is a Fortran90 construct to fix it (one can rename a USE-associated type using the => operator within the USE statement).  

Before I make this change I wanted to give you a heads up and a chance to object.

John


git-svn-id: https://windsvn2.nrel.gov/FAST/branches/CSC/FAST/Registry/Source@90 10acb478-4768-415a-8850-bacdb5912d4d