Changes to address https://wind-dev.nrel.gov/bugzilla/show_bug.cgi?id=486
This is only part-1, have not incorporated Mike's changes yet (see last sentence of note, below).
Modification to ExtrapInterp routines to fix numerical issue when t is large

To solve the numerical issue Tiago reported in the ExtrapInterp routines, I created local time variables, which are just the input t and t_out minus t(1). 

I renamed the subroutine arguments and made t and t_out local variables. See attached code for the first part of the routine. All changes are marked with 
!>>>>>>>> BJJ

The remaining part of the routines can remain the same, but the equations can be simplified because now t(1) = 0. Mike will work on simplifying the equations and send them to you.

git-svn-id: https://windsvn2.nrel.gov/FAST/branches/CSC/FAST/Registry/Source@341 10acb478-4768-415a-8850-bacdb5912d4d