Bug fix in WAMIT2
There was an issue in the 3D and 4D data reading routines.  If a pair of points has to be added at the end to create a step function cutoff, a new frequency is inserted with a value o the highest frequency + EPSILON.  However, the highest frequency and the new frequency would get evaluated incorrectly in a comparison, and the second point would get flagged as not having anything associated with it, as a sparse point.

To fix this, I simply changed EPSILON to 4*EPSILON for the offset for the new point.  This prevents a EqualRealNos comparison from mixing up the two points when creating the mask for points that are populated.


git-svn-id: https://windsvn2.nrel.gov/HydroDyn/trunk@591 c0d87ae3-59f3-4fc3-86f6-af0122200c1e