Bug fix in WAMIT2
Changed how negative indices for force components are handled:
These are now ignored with a warning that the moving reference frame is not handled by HD.  This change required reading that raw data from the file into a temporary array, counting the number of lines that will be kept, then copying those to the RawData3D (or 4D) array for further processing.

Changed the method for populating the frequencies:
Since the 2nd order files may or may not include the zero or infinite frequencies (the second order never do, but for portability to the first order case, this was considered), two points before the start of the frequency data or after it must be added to the array.  These allow for one point at zero (or HUGE), and one close to the lowest/highest frequency in the file (at some Delta away).  This allows for a step function cutoff with the added points having a value of 0.0_ReKi.  The distance Delta has been changed to be 10.0_ReKi*EPSILON in order to give a significant enough distance for the EqualRealNos routine to not confuse the two points.  During the population of the arrays, the index numbers corresponding to the first and last datapoints that were read in from the file (not the added points) are now used.  This eliminates any issues during the population phase (these indices are not stored though since the interpolation routines don't support search bounding at this point).


git-svn-id: https://windsvn2.nrel.gov/HydroDyn/trunk@596 c0d87ae3-59f3-4fc3-86f6-af0122200c1e