Efficiency updates
Removed AFInfo from fcnArgs sent to brent solver; instead pass an AFInfo argument so we don't have to copy/destroy the airfoil table all the time.

Save the residual values computed at the solution region boundaries so that they can be passed to the Brent solver, reducing the number of function evaluations by 2 for each node and time step.

Former-commit-id: 63ec56a7de865d3876ae99b7eba74fe79ce635d5