Airfoil table stores alpha in radians instead of degrees
For consistency (and to avoid multiplying so many things by R2D each node and time step), I'm storing the airfoil tables in radians. It looks like MLB tried this at one point, so I am wondering if there is some numerical tradeoff in the cubic spline interpolation with smaller values of x.

Former-commit-id: 6a71bdad2f36895d8e2aed617366b1c960048136