Removed error log file
Instead of this log file, either (1) pipe the screen output to a file or (2) look at the return code from FAST (the process returns a non-zero value if FAST ended abnormally). 

If you do need a text file of what gets written to the screen, I think it makes more sense to change the NWTC_Library to handle this.