Allowing for new restart capability in OpenFAST - C++ API
While OpenFAST does have a restart capability, some of our work is dependent on the use of a third party turbine controller using the Bladed interface. The Bladed interface could have a restart capability, but is not guaranteed to have one. To work around this limitation, the OpenFAST-C++ API will provide a new capability as shown below:

When the C++ driver starts from time zero, OpenFAST will start normally and write the velocity data the Aerodyn nodes to a file every time step. When the C++ driver restarts from a non-zero time (time > 0), we have two options

1. trueRestart - All components of FAST including the turbine controller are expected to have a checkpoint file at the specified start time. The C++ driver will call OpenFAST to restart from the checkpoint files at the start time and then proceed to step together with OpenFAST for the rest of the simulation.
2. restartDriverInitFAST - The C++ driver will call OpenFAST to start from scratch (time = 0), read the file containing the velocity data at the Aerodyn nodes, and then step OpenFAST upto the start time of the simulation by applying the velocity data at the Aerodyn nodes. Once OpenFAST reaches the start time of the simulation, the C++ driver and OpenFAST will continue to step together normally while continuing to write out the velocity data the Aerodyn nodes at every time step.