Moving supercontroller data into a separate class stored as an object within the OpenFAST - C++ API
The previous OpenFAST - C++  interface stores the states of the supercontroller within a separate supercontroller class with no control over the data. This does not conform to the FASTv8 interface because the supercontroller class cannot correct the supercontroller state once updateStates is called. Hence the OpenFAST - C++ API needs to store all the super controller states along with a duplicate copy for the prediction at the next time and pass it to the supercontroller functions everytime. Finally, when we are sure that enough corrections have been made, the states have to be updated to the next time step. Hence this commit does the following:

1. Move all existing supercontroller states storage into a separate Supercontroller class, create extra storage to store the predicted states at the next time step.
2. Change updateStates and calcOutputs to pass the states into the function.

Any simulation with a supercontroller now needs to only supply a dynamic library with two functions:

1. void sc_updateStates(int nTurbines, int nInputsGlob, std::vector<double> & sc_inputsGlob, int nInputsTurbine, std::vector<double> & sc_inputsTurbine, int nGlobStates, std::vector<double> & globStates_n, std::vector<double> & globStates_np1, int nTurbineStates, std::vector<double> & turbineStates_n, std::vector<double> & turbineStates_np1);
  sc_updateStates_t * sc_updateStates;
2. void sc_calcOutputs(int nTurbines, int nInputsGlob, std::vector<double> & sc_inputsGlob, int nInputsTurbine, std::vector<double> & sc_inputsTurbine, int nGlobStates, std::vector<double> & globStates, int nTurbineStates, std::vector<double> & turbineStates, int nOutputsGlob, std::vector<double> & sc_outputsGlob, int nOutputsTurbine, std::vector<double> & sc_outputsTurbine);