SrvD: replaced y_prev with y in calls to update SrvD inputs
This fixes issue with DLL controller getting constant values for GenTrq and ElecPwr, as reported in https://github.com/OpenFAST/openfast/issues/40. 

If SrvD%y is ever used in another way (e.g., linearization), this method will have to be fixed. Note that it breaks the framework--but the Bladed DLL breaks the framework, too.