Support a CMake-generated Visual Studio solution (#150)
* update the cmake configuration for generating visual studio projects

* bug fix for cross platform compatability in the regression test scripts

* Remove requirement for mkl.h and update default search paths

If the intel c compiler is not installed, this header is not found. Thus, the code is commented out but left in the source for future reference.
Also, search paths for Intel on Windows are included.

* Specify a default installation path for registry

This enables Visual Studio to find the registry executable when it is generated by cmake, and it should have no affect for unix builds.

* Explicitly search for a python executable in pFUnit configuration

The native python search in pFUnit does not find python on Windows, so a working executable can be passed through OpenFAST's CMake.
This commit also removes a deprecated configuration line and includes a bug fix for setting up the unit tests in a cross platform manner.

* Bug fix in setting the preprocessor for multiple compilers

The Intel compiler requires "fpp" while the gnu compiler requires "cpp". Furthermore, this fixes an improper flag character used for Windows command line.

* Add CMake for Windows documentation