TurbSim bug fix for user-specified time-series input files
This affects cases where the number of components specified in the file were less than 3 or if the number of frequencies specified in the input file (through the time histories) were less than the number of requested output frequencies. In these cases, some parts of the velocity spectra were uninitialized or initialized improperly, giving random results.