Fixed undefined behaviour in VariousWaves_Init.
InitOut%WaveDirArr(InitOut%NStepWave2) is not set in case of
multi-directional waves. This leads to undefined behaviour (read of
uninitialized value).

Set this value to 0 as workaround (this generates defined behaviour).
The correct fix is unclear: maybe InitOut%WaveDirArr(0) should be set
to 0 and the index K starting from 1 then???