BeamDyn: Add initial azimuth to driver
Added a way to set the initial rotation separately from the global orientation. This is a flag added to the input file for the driver.

When true, the orientation as the blade root orientation and calculation reference frame orientation.
If the flag is false, it is used only as the blade root orientation (identity matrix for the calculation reference frame orientation).