AD14: Restructure loop to fix optimization bug
Since there is a read and write in the same line, these do-loops shouldnt be optimization by the compiler. However, GNU Fortran version 8+ does optimize anyway leading to a seg fault.

This change removes the read-write operation in the same line by replacing the loops with a more explicit summation.