Solve Option 2 Performance enhancements (#164)
* removed SOLVE_OPTION_1_BEFORE_2 preprocessor directive

* Reworked AdvanceStates and SolveOption2 algorithm

Parts of the SolveOption2 algorithm have been moved to the the AdvanceStates routine so that BeamDyn, InflowWInd, AeroDyn, and ServoDyn use calculated inputs instead of extrapolations for inputs at t+dt. 

This change allows BeamDyn to use calculate rotations and AeroDyn/DBEMT to use calculated velocities, which otherwise would require a correction step. In DBEMT this was a problem with high turbulence and low wind speeds.

* fixed issue for outputting ED mass matrix with preprocessor directive

* HD: remove non-standard tab formatting

* Remove linearization case files from mamual reg test

* Increase the default stack size on Windows

* Update reg test baselines