Fix bug that could cause incorrect augmented mesh in L2-L2 or L2-P load
Line2-to-Line2 or Line2-to-Point load transfers in some instances could have created augmented meshes with nodes in incorrect places, and possibly cause incorrect loads transfer in the process.

Also fixed the mesh debugging code so that if a mesh has scalar quantities, it stores the number of scalars that it is writing to the binary file.