SrvD: revise servodyn main input file parsing to allow passed
NOTE: there is a possible issue with the -fstack_reuse='all' compiler flag on gfortran 9.1.0
with this code!  To demonstrate this bug, comment out the "CurLine=CurLine" in the Failed
function near line 2361.  Without this line, CurLine and ErrStat end up sharing the same
stack location, so ErrStat gets overwritten with each call to Failed!  This may indicate a
potential for other variables to be overwritten by their subroutines during strong
optimization (not certain this exists in other compilers or not).