Fixed indexing bug when setting InitOut data
n the modified code lines, the data on the left side of the equal sign is allocated from 0:NStepWave because the NStepWave element is the ‘wrapped’ value (it ends up being set to the 0th element value) further down in the code.

 

The data on the right side of the equal sign is allocated from 0:NStepWave-1 because it is intermediate data and does not require the ‘wrapped’ element.