SrvD: remove preprocessor directive for method that was abandonned
from @bjonkman regarding the SIMULINK_DirectFeedThrough directive:

That was added when we were deciding how to couple FAST8 with Simulink. If we used the direct feed-through option, we'd have to set the inputs at a different place. In the end, we decided not to go with that coupling method, but I just left the code in there to help remember some of the differences. I feel like there was another place with that directive, though (that's not in the code any more), so maybe we should just get rid of it now.