[BugFix] unallocated AD%y%Rotors when AD14 is used with cpp interface
The FillOutputAry_T routine tried to pass Turbine(iTurb)%AD%y%Rotors(1) to FillOutputAry, but the AD%y%Rotors was not allocated for cases where AD15 was not used.  The logic within FillOutputAry and the variables passed into it have been modified accordingly.

NOTE: this set of routines is only called from the FAST_Library, so this issue would only appear with the cpp interface when AD14 is used (I only know of one Simulink example case that this would affect)