[BugFix] HD: RodrigMat routine in Morison was returning uninitialized junk under certain conditions
The R_I matrix would be returned without setting it to any value if the vector length passed in was 0.  The error handling was also broken.
Since the vector passed (I_n) in would not get set to any value if PropPot was false, and that would leave whatever junk was in memory for R_I which then got used later.