SrvD: move StC inputs/outputs into misc, setup meshes for passing motions/loads to/from SrvD from glue code
This introduces a new set of meshes in SrvD u% and SrvD y% that will be used for passing motions into SrvD and loads from the StC submodule back out.

Next step: add setup passing of StC u% and y% within SrvD.  This includes mapping the SrvD meshes to the StC instances.