HD input parse: new datatype InputFile for holding data from input file
Input file data had originally been stored in InitInp so that it could be passed in directly.  However, we are now migrating to a method of passing all the input file data as datatype FileInfo (basically a string array with some meta-data).  This simplifies setting up to pass input file information from a different language (such as python through a c-bound library interface to HD -- which is needed for an ATLANTIS project).

It is a little bit messier than I like at the moment -- the InputFile datatype contains some WAMIT, Waves, and other InitInput types.  But right now I don't see a simpler option that won't require lots of work.