Add Fortran pointers for multi-dimensional arrays
If the user specifies the array dimensions via {*} instead of {:} a Fortran pointer to an array is generated instead of an allocatable array.  If the -ccode option is used, then {*} results  in the same output as {:} for compatibility with the corresponding C code.  The user can only generate these Fortran pointers in the InitInput, InitOutput and Parameter data structures.  The Registry will exit with an error if the user attempts to generate Fortran pointers in the other Framework data structures.