Bug Fix: Non-zero MSL2SWL was incorrectly translating Morison Meshes
Internally, the HydroDyn calculations use the Still Water Line as the Z=0 reference, and this was being applied to the Morison input and output mesh data.  This was a bug.  The mesh data should have been referenced to the Mean Sea Level, which corresponds to OpenFAST's inertial coordinate system.

This code, corrects the node positions for the Morison input/output meshes.

Addionally, the Marine growth depth values in the HD input file were being shifting using the MSL2SWL parameter, but this was also a bug.  The new code no longer shifts these input file depth values.

Internally, this change requires the Morison module's  CalcOutput routine to shift the input mesh positions to the Still Water Line reference system for the hydrodynamics equations.