SD: bug fix, rigid assembly with interface, order mattered
For a rigid assembly that involves an interface joint, the joint selected as "leader"
is always the interface node. There was a bug when the interface node
was somehow futher down the list of joints as the other joints involved
in the assembly.

The list INodesID returned by RAElimination was correct, but after this
function call, there was a need to directly address the other joints and
change the "iNode" of the for loop so that it points to the leader Node.
The variable "iNodeSel" was introduced to account for this.