Fix Bug in FAST.Farm Causing Wake Bounce-Back (#839)
For wake planes that are skewed relative to domain boundaries (e.g., due to nacelle-yaw), the weighting used in the spatial averaging to calculate the meandering velocity of the plane when Mod_Meander = 3 could end up being negative as wake planes were leaving the domain.  But the code always assumed that the weighting was positive.  This bug fixes that.  This solves the issue reported by on our forum dated July 06, 2021: https://wind.nrel.gov/forum/wind/viewtopic.php?f=13&t=2117&start=225.