IfW grid exceed: algorithm for extrapolating from tower below grid
This gives a reasonably continuous function (not smooth) from the tower out beyond the +/-Y edge of the grid.  There is some physics that we can't account for simply because the information doesn't exist, but this should at least allow wake evolution.  If weirdness occurs, extend grid to ground.

One missing piece at the moment is the summation of NGe should be ==1 for the region between 2*FFYHWid < abs(y) < FFYHWid, but it is not.  Since the plots show continuity in value (not in derivatives), it is assumed there is some book-keeping incorrect with the NGe(5) and NGe(10) terms which get multiplied by zero.  So something isn't exactly right in the derivation, but the result is still reasonable (though potentially incorrect).