[bugfix] SrvD: fix inconsistency in initial pos/orient for StC
The reference position, reference orientation, translational displacement, and orientation at init are now passed to SrvD for all the StC instances.  It is possible to do this with less info since some orientations could be assumed at init, but for throroughness we pass all of it to ServoDyn to use in initializing the StC instances.  This in some ways future-proofs for when arbitrary initial orientations are added to the code (if ever)