nwtc_io bugfix: prevent overrunning end of FileInfoType
If the last line does not have a carriage return, the file may be passed with no NULL at the end of the last line.  If this happens and the last line starts with "END", the routine for finding the outlist may overstep the end of hte FileInfoType and read garbage (or segfault) -- at the very least give nasty errors from whatever code called it to parse the FileInfoType into the file structure.  This fixes this issue.