[BugFix] BeamDyn nodal outputs occasionally segfaulted
The cause was an index that never got set.  So this could result in
attempting to retrieve a value from the output mapping array using an
out of bounds index (sometimes a very large negative number).  This
never appeared in gcc compiled versions, but occasionally would appear
in certain intel compiled versions (others initialized the value
apparently).