MDv2: More strict parsing of Type/BodyID for pinned Rods
Consider a Rod only as pinned if input is exactly "PINNED" or "PIN". Anything else now results in fatal error.
Thus, "pinnedx", "pinxned" or similar which used to be interpreted as a pinned rod now result in an error.
Also, if BodyID is followed by any characters it will now result in an error. Before, those characters were ignored and the Rod interpreted as fixed.