NWTC Lib / Version Info: move all version info code to VersionInfo module
moved all code that was in NWTC_IO that called the version info module into the version info module. This allows us to rebuild the version info stuff without having to rebuild the NWTC library.

Also moved some of the more generic GetVersion functions from FAST and SeaState into the version info module so we won't have to duplicate code with every driver that wants to use it.