Fix Time-Step Delay in Super Controller within FAST.Farm
The transfer within FAST.Farm of to_SC, which is an output from OpenFAST and an input to the super controller, previously happened after the call to the super controller CalcOutput, which caused a one time-step delay. This change moves the transfer before the call to super controller CalcOutput to eliminate the delay. The original implementation correctly followed the FAST.Farm glue code implementation plan, but it is unclear why the plan was written this way. The FAST.Farm user documentation in readthedocs is already consistent with this change, so, the implementation and user documentation are now consistent.  This issue was discussed in Dec '21 on our forum: https://forums.nrel.gov/t/super-controller-link-with-turbine-controller/2587/19.