Added the AMMod entry to the 'STRIP THEORY OPTIONS' section of the HydroDyn input file.
When AMMod = 0, the distributed added-mass force is always and only computed on nodes with reference position below SWL. When AMMod = 1, the distributed added-mass force is computed up to the instantaneous free surface with load smoothing. As with the fluid inertia force and drag force with wave stretching, the displaced z-position is used to determine whether a node is below the free surface when WaveDisp = 1 and the undisplaced z-position is used when WaveDisp = 0 when AMMod = 1.  Note that when WaveMod = 0 or 6 or when WaveStMod = 0 in SeaState, AMMod is overwritten to 0. In other words, AMMod = 1 only takes effect when wave stretching is enabled.