r-test: missing test case after merge of #1217
Test case #1217 included test case hd_5MW_OC4Jckt_DLL_WTurb_WavesIrr_MGrowth, but the dev branch on r-test did not include this.  In the previous commit, I mistakenly changed to OpenFAST/dev pointer to point to the r-test/dev which did not include the case.  This effectively removed it.  This commit along with the commit on r-test/dev puts this case back into the regression test suite.