SD: SumPrint line in input file had name of SSSum
If there was a file parsing error in JOINT ADDITIONAL CONCENTRATED MASSES section, the `SumPrint` line in the input file could have a read error.  This results in an error printed to the screen about `SSSum` not found -- but the input files all say `SumPrint`.

Another option instead of changing this is to update all the SD input files to have `SSSum`.  However, this would be nonstandard with the rest of our input files