[BugFix] rtd configuration changes broke our build
I have a suspicion that the argument parsing in python or some package in it changed (I can reproduce the issue locally).

There is an arguement in `CMakeListst.txt` for the `sphinx-html` command that indicates where the `conf.py` is located (`-c ${CMAKE_CURRENT_BINARY_DIR}`). This is at the end of the command for `sphinx-html`.

The argument parsing in `conf.py` is assuming this last argument doesn't count as it finds the arguments for the path to `doxygen` and `openfast.tag`.  Something somewhere appears to have changed this behavior.

Removing this last argument restores the ability for me to build locally and for rtd to build and really has no impact as conf.py gets copied to the build dir.

Still have no clue what changed at rtd to cause this -- maybe a newer python?