[BugFix] Doxygen builds failing on rtd, and locally (#1442)
* Disable doxygen on readthedocs as a temporary measure (does not really address #1423)
* Fix local compiling of the docs when doxygen is enabled:

**Description:**
Something in RTD changed in January 2023 causing our documentation to fail if doxygen is enabled. This is due to an issue with the paths to doxygen. The change occurred sometime after the 3.4.0 release.

* v3.4.0 (main, e8ec53f) docs build on Jan 11, https://readthedocs.org/projects/openfast/builds/19142412/ --> pass
* v3.4.0 (main, e8ec53f) docs build on Jan 30, https://readthedocs.org/projects/openfast/builds/19322116/ --> fail

Exactly what changed is a bit unknown: looking through the logs for both yielded very little useful information. A handful of python packages had minor updates, but nothing major that we would expect to change the results. Also whatever changed also affects local builds (only for updated python packages).

For this fix, doxygen has been turned off on RTD (we don't think anyone uses it). Doxygen is still built locally and should work on most systems still.