AWAE: bugfix -- IfW needs HubPosition and rotor rad
InflowWind always calculates the disk average velocity and the velocity at the hub. AWAE was not setting the HubPosition or the RotorRad variables for IfW.  So depending on the compiler and optimizations, IfW would sometimes fail in AWAE when the wrong random garbage was in those memory locations (interestingly the GH compiler did not trip on this issue).