BeamDyn missing velocity init in BD_InitAcc
When using BeamDyn from OpenFAST, there was a convergence error during
the initial correction steps which was not present when using the
driver. After investigating, it turns out that the velocity in
MiscVars%vvv was not being reinitialized in BD_InitACC during each
correction and was accumulating offsets which caused the model to
not converge. This commit adds the call to initialize the velocity which
is present during the dynamic solve.