IfW: fix incorrect use of p%RefPosition
When IfW was refactored, RefPosition was moved inside the Flow Field
derived type; however, the code that set this value was not updated
properly so RefPosition as always zero. This was missed by regression
testing because the wind was not rotated. A new ifw_uniform test has
been added to catch this in the future.