Use reg libs in CMake, build FAST_SFunc in CMake
This commit undoes most of the changes in PR #1010 which introduced
object libraries into the build system to produce a statically linked
MEX file for Simulink. The Object Libraries were somewhat complex and
didn't behave the same as normal libraries. While trying to debug a
compilation issue on an M1 mac the CMake `matlab_add_mex` function was
discovered which could build a static mex file within CMake. The build
system was reworked to use this feature.