Move OpenMP to main CMakeLists.txt
There was some duplication in processing of OpenMP flags between
FAST.Farm and openfastcpp which caused compile issues. Finding the
package and adding flags were moved to the main CMakeLists.txt. This also sets
the LINKER_LANGUAGE to Fortran for openfastlib and fixes a typo in the
VersionInfo CMake file.