Port openfast-registry to C++ and change pointers
This commit contains a complete port of the openfast-registry to C++.
The new registry doesn't support the dimspecs used in AD14 so
they were replaced in the input file. This module will be removed soon
and this feature is not used by other modules so it didn't seem
necessary to keep this feature.

AeroDyn_Registry.txt had a malformed line which was acceptable to the
previous parser, but didn't work for the new one so the input file was
updated.

The registry implements specifying a target and a regular pointer
pointer on the field name using & and * respectively. The registry
treats TARGETs (&) the same as allocatable fields wrt pack/unpack/copy.
Essentially, target's own their memory. POINTERS (*) do not own their
memory and are not packed/unpacked. The pointer association is
maintained by the copy function.