Escaped " in AeroDyn & AeroAcoustics Registry
The new registry parses files and keeps quoted strings intact.
There were a couple of instances where "default" appeared inside
a quoted string which was then interpreted as separate fields.
This can be avoided by escaping the double quote as \" so the
parser won't end the string.

There was also an extra dash on line 320 in AeroDyn_Registry.txt
which caused the description to be printed as a unit.