Use pointers to couple InflowWind and FAST.Farm
This commit is the initial attempt to use pointers to couple the wind
velocities generated by AWAE in FAST.Farm to the InflowWind
FlowField which is used by AeroDyn to calculate the wind velocities.
The Vdist_High data in AWAE has been changed to a pointer and
FlowField%Grid4D%Vel is associated with this data to alleviate the
need to copy the data between the modules.