[BugFix] Lidar: don't allocate already allocated array in SrvD
The input arrays for Lidar to ServoDyn get allocated in ServoDyn.f90, and again in BladedInterface_EX.f90.  This double allocation causes issues for many compilers.  The logic behind all this really should be looked at in more detail, but since the Lidar module in InflowWind will be superceeded by LidarSim in release 4.0.0, I'm not going to sort the logic right now.

Affected arrays:
    u%LidSpeed
    u%MsrPositionsX
    u%MsrPositionsY
    u%MsrPositionsZ

Thanks  to @foxton9 for reporting the issue, and proposing the fix!