Update Registry Pack/Unpack to support Pointers
This commit significantly modifies the OpenFAST Registry code which
generated the Pack and Unpack routines so that it can support packing
and unpacking pointers without loosing their associations. This was
accomplished by creating a PackBuffer derived type which stores
a buffer of bytes and a pointer index. The byte buffer replaces the
DbKiBuf, IntKiBuf, and ReKiBuf and packs all data as bytes (int8) using
the `transfer` intrinsic. As each pointer is encountered in any
structure, `c_loc` is used to get the address and the pointer index
in PackBuffer is searched. If the pointer is found (already encountered
in another field), then the exisitng index is saved into the buffer;
otherwise, the pointer is added to the index and the new index is saved
into the buffer. If the pointer wasn't in the index, then the associated
data is saved to the buffer. Unpacking is the reverse of this process.