FF: fix bug with Mod_AmbWind=3 -- hub point wasn't shifted.
We added the hub point to AWAE at the center of the grid now that IfW calculates the disk average velocity every time IfW_CalcOutput gets called. But this point needs to be shifted by the turbine position so it is in the box.

This will all change in 4.0.0 when we switch to pointers for all this and don't call IfW_CalcOutput directly anymore.