Resolve BeamDyn initial strain for rotated blade
In the BeamDyn driver, and likely in the glue-code, if the blade was
initially rotated then an initial strain was introduced which produced
forces when no loads or gravity had been applied. The cause was
traced to the calculation of R0 (initial rotation) at the quadrature
points which was interpolated via the shape functions from the
element nodal rotations. The shape function interpolation is not
exact and introduced strains on the order of 1e-7, but with the large
stiffness of the blade, forces of 6200 N could be produced at the root
for the IEA 15MW blade. This was resolved by directly calculating
the quadrature point initial rotations in uu0(4:6) via interpolated
tangent and twist, which are well behaved. This was only a problem for
non-straight blades such as the IEA 15MW.

This commit also fixes a bug where the parameter UsePitchAct
wasn't initialized from the input file until after it was used in
SetOutParam. The initialization has been moved to SetParameters.