Fix missing symbols for Simulink and  LAPACK build
When building the Simulink MEX file with CMake on Linux, the resulting
shared library had unresolved symbols. This was due to the order in
which the static libraries were specified in simulink/CMakeLists.txt.
The GCC linker does one pass over the libraries and notes unresolved
dependencies which can only be resolved by the succeeding libraries.
This means that the most generic libs (like NWTC-Library) need to
appear at the end of the list.

Instead of leaving it up to the user to get the order right, the
simulink/CMakeLists.txt file has been updated to link the libraries
twice.

The same issue occured when using USE_LOCAL_STATIC_LAPACK
in that the BLAS libraries were linked before LAPACK so there
were unresolved symbols. This has been fixed by reversing the order.

Also, the OpenFOAM libraries have been marked as STATIC.