[BugFix] change time handling to double precision for G4D timestep index
The time handling for G4D time index finding was a mix of single and double precision.  This could lead to finding the wrong index in a G4D wind grid passed from FF to an OpenFAST turbine, in which case the index might be outside the G4D box that was passed.  We think this can occur with a long simulation of 4,200 s, with an OpenFAST DT of 0.0378 s, which resulted in a time index out of bounds at time 4132 seconds.